将activity中布局分成moudle添加

+++ GoogleStore/src/main/java/com/ybbbi/googlestore/bean/AppInfo.java
@@ -17,7 +17,7 @@ public class AppInfo {
     public String name;
     public String packageName;
     public long size;
-    public long stars;
+    public float stars;
     public String des;
     public ArrayList<String> screen;
     public ArrayList<safeInfo> safe;
